use uint32 instead of u_long

+++ pppd/magic.h
@@ -16,9 +16,9 @@
  * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
  * WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR PURPOSE.
  *
- * $Id: magic.h,v 1.1 1993/11/11 03:54:25 paulus Exp $
+ * $Id: magic.h,v 1.2 1994/09/01 00:25:35 paulus Exp $
  */
 #include "args.h"
 
 void magic_init __ARGS((void));	/* Initialize the magic number generator */
-u_long magic __ARGS((void));	/* Returns the next magic number */
+uint32 magic __ARGS((void));	/* Returns the next magic number */
